pyproject.toml
[project]
name = "rankings-uploader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.1",
    "curl-cffi>=0.14.0",
    "python-dotenv>=1.2.1",
    "selectolax>=0.4.6",
]
